requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
bcrypt
email-validator
python-multipart
pymysql
PyJWT
python-jose[cryptography]